requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
python-dotenv
requests
numpy==1.26.4
sqlalchemy==2.0.29
pydantic==2.6.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
google-auth==2.34.0
python-jose[cryptography]==3.3.0
pydantic[email]
google-cloud-vision
google-cloud-secret-manager
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary>=2.9.6
google-cloud-storage==2.18.2
